+++ CHANGELOG.md
@@ -60,6 +60,7 @@ hi
 - Peonies now generate in Birch Forest biomes.
 - Lilacs now generate in Birch Forest biomes.
 - Termite Mounds no longer have a chance to generate with a disk of Sand beneath them.
+- Jellyfish Caves have been renamed to Mesoglea Caves.
 - Updated Ukranian translation thanks to [unroman](https://github.com/FrozenBlock/WilderWild/pull/413)!
 
 # Bug Fixes